# backend/docs/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reddit Signal Scanner",
    "description": "Reddit Signal Scanner API (Phase 1/2 implementation)",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create a new account and receive an access token",
        "operationId": "register_user_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Authenticate existing user and receive an access token",
        "operationId": "login_user_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get current authenticated user",
        "operationId": "get_current_user_me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/status/{task_id}": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "获取任务状态（缓存优先）",
        "operationId": "get_task_status_api_status__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/stats": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "获取任务队列统计信息",
        "description": "Return aggregate Celery worker statistics to support operational dashboards.",
        "operationId": "get_task_stats_api_tasks_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/insights/{task_id}": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "获取指定任务的洞察卡片列表",
        "description": "根据任务 ID 返回结构化洞察卡片，支持置信度和子版块过滤。",
        "operationId": "list_insights_by_task_api_insights__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "最小置信度（0-1），用于过滤洞察卡片",
              "title": "Min Confidence"
            },
            "description": "最小置信度（0-1），用于过滤洞察卡片"
          },
          {
            "name": "subreddit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按子版块过滤，支持精确匹配（区分大小写）",
              "title": "Subreddit"
            },
            "description": "按子版块过滤，支持精确匹配（区分大小写）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 10,
              "title": "Limit"
            },
            "description": "每页数量"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightCardListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/insights/card/{insight_id}": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "获取单个洞察卡片",
        "description": "根据洞察卡片 ID 获取详细信息",
        "operationId": "get_insight_api_insights_card__insight_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Insight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthTokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "expires_at",
          "user"
        ],
        "title": "AuthTokenResponse"
      },
      "AuthUser": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "membership_level": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MembershipLevel"
              }
            ],
            "default": "free"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "AuthUser"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InsightCardListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "总数"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InsightCardResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "洞察卡片列表"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "InsightCardListResponse",
        "description": "洞察卡片列表响应 Schema"
      },
      "InsightCardResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "洞察卡片 ID"
          },
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id",
            "description": "关联的分析任务 ID"
          },
          "title": {
            "type": "string",
            "maxLength": 500,
            "title": "Title",
            "description": "洞察卡片标题"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "洞察摘要"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "置信度分数 (0.0-1.0)"
          },
          "time_window": {
            "type": "string",
            "title": "Time Window",
            "description": "时间窗口说明，例如“过去 30 天”"
          },
          "time_window_days": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Time Window Days",
            "description": "时间窗口（天数）"
          },
          "subreddits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subreddits",
            "description": "相关子版块列表"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence",
            "description": "证据列表"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "title",
          "summary",
          "confidence",
          "time_window",
          "time_window_days",
          "created_at",
          "updated_at"
        ],
        "title": "InsightCardResponse",
        "description": "洞察卡片响应 Schema\n\n注意：不使用 from_attributes=True，因为我们在 Service 层手动构造\n避免 Pydantic 递归访问 ORM 关系导致 RecursionError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "MembershipLevel": {
        "type": "string",
        "enum": [
          "free",
          "pro",
          "enterprise"
        ],
        "title": "MembershipLevel",
        "description": "Enumerated membership tiers available to end users."
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "membership_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MembershipLevel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest"
      },
...
      "TaskStatsResponse": {
        "properties": {
          "active_workers": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Active Workers",
            "description": "Number of active Celery workers"
          },
          "active_tasks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Active Tasks",
            "description": "Count of tasks currently executing"
          },
          "reserved_tasks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Reserved Tasks",
            "description": "Count of tasks reserved by workers"
          },
          "scheduled_tasks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Scheduled Tasks",
            "description": "Count of tasks scheduled for later execution"
          },
          "total_tasks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Tasks",
            "description": "Aggregate of active, reserved, and scheduled tasks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "active_workers",
          "active_tasks",
          "reserved_tasks",
          "scheduled_tasks",
          "total_tasks"
        ],
        "title": "TaskStatsResponse",
        "description": "Celery task queue statistics."
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "completed",
          "failed"
        ],
        "title": "TaskStatus"
      },
      "TaskStatusSnapshot": {
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "percentage": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percentage"
          },
          "current_step": {
            "type": "string",
            "title": "Current Step"
          },
          "sse_endpoint": {
            "type": "string",
            "title": "Sse Endpoint"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count",
            "default": 0
          },
          "failure_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Category"
          },
          "last_retry_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Retry At"
          },
          "dead_letter_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dead Letter At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "status",
          "progress",
          "message",
          "percentage",
          "current_step",
          "sse_endpoint",
          "updated_at"
        ],
        "title": "TaskStatusSnapshot"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}